modify original qrcode package from metian9288
