+++ .idea/modules.xml
@@ -2,9 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/flutter_qr_reader.iml" filepath="$PROJECT_DIR$/flutter_qr_reader.iml" />
       <module fileurl="file://$PROJECT_DIR$/android/flutter_qr_reader_android.iml" filepath="$PROJECT_DIR$/android/flutter_qr_reader_android.iml" />
       <module fileurl="file://$PROJECT_DIR$/example/android/flutter_qr_reader_example_android.iml" filepath="$PROJECT_DIR$/example/android/flutter_qr_reader_example_android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/super_qr_reader.iml" filepath="$PROJECT_DIR$/super_qr_reader.iml" />
     </modules>
   </component>
 </project>
