+++ CHANGELOG.md
@@ -1,28 +1,3 @@
-## 0.0.1
+## 3-25-2020
 
-## 0.0.2
-
-增加内置UI
-
-## 0.0.3
-
-## 0.0.4
-
-## 0.0.5
-
-## 1.0.0
-
-## 1.0.0+1
-
-## 1.0.1
-
-修改扫码后的操作
-
-## 1.0.2
-
-降低zxing版本至3.3.3
-
-## 1.0.3
-
-解决scanBoxRatio无效问题
-
+0.0.1: modify original library to meet our business requirement.
