+++ example/android/gradle.properties
@@ -2,3 +2,4 @@ android.enableJetifier=true
 android.useAndroidX=true
 org.gradle.jvmargs=-Xmx1536M
 
+android.enableR8=true
