+++ lib/qrcode_reader_view.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'flutter_qr_reader.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:permission_handler/permission_handler.dart';
+
+import './qrcode_reader_controller.dart';
 
 /// 使用前需已经获取相关权限
 /// Relevant privileges must be obtained before use
@@ -13,13 +15,13 @@ class QrcodeReaderView extends StatefulWidget {
   final Color boxLineColor;
   final Widget helpWidget;
   QrcodeReaderView({
-                     Key key,
-                     @required this.onScan,
-                     this.headerWidget,
-                     this.boxLineColor = Colors.cyanAccent,
-                     this.helpWidget,
-                     this.scanBoxRatio = 0.85,
-                   }) : super(key: key);
+    Key key,
+    @required this.onScan,
+    this.headerWidget,
+    this.boxLineColor = Colors.cyanAccent,
+    this.helpWidget,
+    this.scanBoxRatio = 0.85,
+  }) : super(key: key);
 
   @override
   QrcodeReaderViewState createState() => new QrcodeReaderViewState();
@@ -31,7 +33,7 @@ class QrcodeReaderView extends StatefulWidget {
 /// qrViewKey.currentState.startScan();
 /// ```
 class QrcodeReaderViewState extends State<QrcodeReaderView>
-        with TickerProviderStateMixin {
+    with TickerProviderStateMixin {
   QrReaderViewController _controller;
   AnimationController _animationController;
   bool openFlashlight;
@@ -46,7 +48,7 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
   void _initAnimation() {
     setState(() {
       _animationController = AnimationController(
-              vsync: this, duration: Duration(milliseconds: 1000));
+          vsync: this, duration: Duration(milliseconds: 1000));
     });
     _animationController
       ..addListener(_upState)
@@ -108,14 +110,20 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
 
   Future _scanImage() async {
     stopScan();
-    var image = await ImagePicker.pickImage(source: ImageSource.gallery);
-    if (image == null) {
+    Map<PermissionGroup, PermissionStatus> permissions =
+        await PermissionHandler().requestPermissions([PermissionGroup.storage]);
+    if (permissions[PermissionGroup.storage] == PermissionStatus.granted) {
+      print('entering picking image');
+      var image = await ImagePicker.pickImage(source: ImageSource.gallery);
+      if (image == null) {
+        startScan();
+        return;
+      }
+      final rest = await FlutterQrReader.imgScan(image);
+      await widget.onScan(rest);
+    } else {
       startScan();
-      return;
     }
-    final rest = await FlutterQrReader.imgScan(image);
-    await widget.onScan(rest);
-    startScan();
   }
 
   @override
@@ -162,7 +170,7 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
                   boxLineColor: widget.boxLineColor,
                   animationValue: _animationController?.value ?? 0,
                   isForward:
-                  _animationController?.status == AnimationStatus.forward,
+                      _animationController?.status == AnimationStatus.forward,
                 ),
                 child: SizedBox(
                   width: qrScanSize,
@@ -172,8 +180,8 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
             ),
             Positioned(
               top: (constraints.maxHeight - qrScanSize) * 0.333333 +
-                      qrScanSize +
-                      24,
+                  qrScanSize +
+                  24,
               width: constraints.maxWidth,
               child: Align(
                 alignment: Alignment.center,
@@ -185,9 +193,9 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
             ),
             Positioned(
               top: (constraints.maxHeight - qrScanSize) * 0.333333 +
-                      qrScanSize -
-                      12 -
-                      35,
+                  qrScanSize -
+                  12 -
+                  35,
               width: constraints.maxWidth,
               child: Align(
                 alignment: Alignment.center,
@@ -201,12 +209,14 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
             Positioned(
               width: constraints.maxWidth,
               bottom: constraints.maxHeight == mediaQuery.size.height
-                      ? 12 + mediaQuery.padding.top
-                      : 12,
+                  ? 12 + mediaQuery.padding.top
+                  : 12,
               child: Row(
                 crossAxisAlignment: CrossAxisAlignment.center,
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: <Widget>[
+                  SizedBox(width: 45, height: 45),
+                  SizedBox(width: 80, height: 80),
                   GestureDetector(
                     behavior: HitTestBehavior.translucent,
                     onTap: _scanImage,
@@ -223,23 +233,6 @@ class QrcodeReaderViewState extends State<QrcodeReaderView>
                       ),
                     ),
                   ),
-                  Container(
-                    width: 80,
-                    height: 80,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.all(Radius.circular(40)),
-                      border: Border.all(color: Colors.white30, width: 12),
-                    ),
-                    alignment: Alignment.center,
-                    child: Image.asset(
-                      "assets/tool_qrcode.png",
-                      package: "flutter_qr_reader",
-                      width: 35,
-                      height: 35,
-                      color: Colors.white54,
-                    ),
-                  ),
-                  SizedBox(width: 45, height: 45),
                 ],
               ),
             )
@@ -262,11 +255,11 @@ class QrScanBoxPainter extends CustomPainter {
   final Color boxLineColor;
 
   QrScanBoxPainter(
-          {@required this.animationValue,
-            @required this.isForward,
-            this.boxLineColor})
-          : assert(animationValue != null),
-            assert(isForward != null);
+      {@required this.animationValue,
+      @required this.isForward,
+      this.boxLineColor})
+      : assert(animationValue != null),
+        assert(isForward != null);
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -299,7 +292,7 @@ class QrScanBoxPainter extends CustomPainter {
     path.moveTo(size.width, size.height - 50);
     path.lineTo(size.width, size.height - 12);
     path.quadraticBezierTo(
-            size.width, size.height, size.width - 12, size.height);
+        size.width, size.height, size.width - 12, size.height);
     path.lineTo(size.width - 50, size.height);
     // leftBottom
     path.moveTo(50, size.height);
@@ -310,7 +303,7 @@ class QrScanBoxPainter extends CustomPainter {
     canvas.drawPath(path, borderPaint);
 
     canvas.clipRRect(
-            BorderRadius.all(Radius.circular(12)).toRRect(Offset.zero & size));
+        BorderRadius.all(Radius.circular(12)).toRRect(Offset.zero & size));
 
     // 绘制横向网格
     final linePaint = Paint();
@@ -346,9 +339,9 @@ class QrScanBoxPainter extends CustomPainter {
 
   @override
   bool shouldRepaint(QrScanBoxPainter oldDelegate) =>
-          animationValue != oldDelegate.animationValue;
+      animationValue != oldDelegate.animationValue;
 
   @override
   bool shouldRebuildSemantics(QrScanBoxPainter oldDelegate) =>
-          animationValue != oldDelegate.animationValue;
+      animationValue != oldDelegate.animationValue;
 }
